+++ rbx_dom_weak/CHANGELOG.md
@@ -1,6 +1,8 @@
 # rbx\_dom\_weak Changelog
 
 ## Unreleased
+* `RbxValueType` now derives `Serialize` and `Deserialize`
+* Added support for `BrickColor`
 
 ## 1.4.0 (2019-03-27)
 * Added support for new types: